+++ src/nncf/common/quantization/quantizer_setup.py
@@ -21,11 +21,11 @@ from nncf.common.logging import nncf_logger
 from nncf.common.quantization.structs import NonWeightQuantizerId
 from nncf.common.quantization.structs import QuantizationScheme as QuantizationMode
 from nncf.common.quantization.structs import QuantizerConfig
+from nncf.common.quantization.structs import TypedQuantizerConfig
 from nncf.common.quantization.structs import UnifiedScaleType
 from nncf.common.quantization.structs import WeightQuantizerId
 from nncf.common.stateful_classes_registry import CommonStatefulClassesRegistry
 from nncf.config.schemata.defaults import QUANTIZATION_NARROW_RANGE
-from nncf.experimental.quantization.structs import ExtendedQuantizerConfig
 
 QuantizationPointId = int
 
@@ -195,14 +195,14 @@ class SingleConfigQuantizationPoint(QuantizationPointBase):
         insertion_point_cls = CommonStatefulClassesRegistry.get_registered_class(insertion_point_cls_name)
         insertion_point = insertion_point_cls.from_state(state[cls._state_names.INSERTION_POINT])  # type: ignore
         qconfig_state = state[cls._state_names.QCONFIG]
-        # Need to instantiate ExtendedQuantizerConfig
+        # Need to instantiate TypedQuantizerConfig
         # to support additional fields used by ExecuTorch-specific quantizer configs.
         # TODO (dlyakhov): Refactor and generalize quantizer config deserialization to cleanly handle both
         # standard and extended formats without relying on manual key comparison (ticket 170078).
         if QuantizerConfig().__dict__.keys() == qconfig_state.keys():
             qconfig = QuantizerConfig.from_state(qconfig_state)
         else:
-            qconfig = ExtendedQuantizerConfig.from_state(qconfig_state)
+            qconfig = TypedQuantizerConfig.from_state(qconfig_state)
 
         kwargs = {
             cls._state_names.INSERTION_POINT: insertion_point,
