+++ src/nncf/quantization/algorithms/min_max/torch_fx_backend.py
@@ -25,9 +25,9 @@ from nncf.common.hardware.config import HWConfig
 from nncf.common.quantization.quantizer_propagation.structs import QuantizationTrait
 from nncf.common.quantization.structs import QuantizationScheme
 from nncf.common.quantization.structs import QuantizerConfig
+from nncf.common.quantization.structs import TypedQuantizerConfig
 from nncf.experimental.common.tensor_statistics.collectors import REDUCERS_MAP
 from nncf.experimental.common.tensor_statistics.collectors import TensorReducerBase
-from nncf.experimental.quantization.structs import ExtendedQuantizerConfig
 from nncf.experimental.torch.fx.commands import FXApplyTransformationCommand
 from nncf.experimental.torch.fx.model_utils import get_target_point
 from nncf.experimental.torch.fx.transformations import qdq_insertion_transformation_builder
@@ -195,7 +195,7 @@ class FXMinMaxAlgoBackend(MinMaxAlgoBackend):
     ) -> FakeQuantize:
         per_channel = quantizer_config.per_channel
         dtype = None
-        if isinstance(quantizer_config, ExtendedQuantizerConfig):
+        if isinstance(quantizer_config, TypedQuantizerConfig):
             dtype = quantizer_config.dest_dtype
 
             if dtype not in [TensorDataType.int8, TensorDataType.uint8]:
